updating gpl2 name identifiers list for WP themes
adding this identifier as many theme authors use underscores as a starter project, and many WordPress themes use "GNU General Public License v2 or later" as a license identifier in `readme.txt` and/or `style.css`.  As many people use this identifier, it's become a commonly known way to identify GPL-2.0-or-later in the WordPress theme community.